Tests were written.

+++ list_api/Models/ListProduct.cs
@@ -6,7 +6,7 @@ namespace list_api.Models {
 		public int IDList { get; set; }
 		public int IDProduct { get; set; }
 		public int Quantity { get; set; }
-		public double Cost { get; set; }
+		public int Cost { get; set; }
 		public List? List { get; set; }
 		public Product? Product { get; set; }
 	}
